Qa/snackbar (#167)
* change: rework snackbar content using slots

Instead of explicitly defining buttons inside the snackbar component, we're going to use slots.

- removes specific subcomponent color props

* pass slots properly

related to #166
- Add slots for: badge, text, button, iconButton

* add fx for handling font scaling

plus sidequest: explicitly define props in Text component

* nest snackbar content in a card

has built-in scalefactoring, anyway.

* sidequest: add optional scrim layer to badge

useful to have a scrim layer with customizable opacity for scenarios like this

* debug button color passing

we want an outline variant for the snackbar buttons and it's requiring manual modifiers defined in the snackbar's index.tsx.

* remove other fontclass options, force to always be body

* fully remove fontClass prop

* sidequest: add stateLayer override prop to button

there will be scenarios where you need to override the button's default state layer handling. this prop, if passed, will supercede the component's native state-layer handling.

* change: replace badge with icon

- replaces badge with an icon, badge was unnecessary and making things look cluttered
- manual override icon positioning using the slot selector in snackbar.css

* color tests passed

also applies minor css changes to spacing and icon sizing

* finished dynamic color setup

minor css corrections: dynamic apply margin-right to the snackbar text if icon is present but the button is not

* dynamic apply m-right if icon present but no button

* snackbar qa complete

* housekeeping: address eslint warnings