Qa/sticker (#169)
* set up staging

* document params

* sidequest: Create debugUtils.ts

After doing QA on nearly all components, it's clear we need a utility file full of arrays of commonly used variables we can import whenever we need to loop through them to test many many permutations of something.

* pass color props as classes

* upgrade getOnToken

now you can pass it any LkColor and it will find a corresponding complement, even if it's not a literal "on" token.

* finish sticker QA