Qa/text input (#172)
* create staging environment

* write: text input component

the original component was completely missing so I just wrote a new one.

* sidequest: change state layer opacity handling

make it more subtle

* test: prop passing

* add labelBackgroundColor prop

necessary so that when label position is on input, and input is focused, it will cover the outline of the input properly

* add CSS: on-field-with-value-set

when labelPosition is on input, this prop ensures that it stays at the top when value has been set on the field, and returns to on-input if that value is erased.

* optimize label spacing

* sidequest: change outline tone

outline token was being rendered at tones._90 instead of tones._60. fixed.

* fix spacing

* lint: remove unused imports

* sidequest: add icon-offset class to icon's base css

realized i'd need it in places besides the button component.

also realized we need to preface literally every custom attribute lk- with data-. how did we not catch that